requirements.txt
Flask==3.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
python-dateutil==2.8.2
requests==2.31.0
